Level 1 & Menu UI
Level 1 is now fully complete with the player able to die and restart the level, as well as putting a finishing point of the level. UI is done as well, giving the user ability to see how to play, start the game, and when they finish or die in the game, they can restart the game or go back to main menu.